Create helpers to calculate project progress and project status
